requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
openai>=1.3.5
python-dotenv>=1.0.1
requests>=2.31.0
gunicorn>=21.2.0  # Needed for production deployment